Fix
SpriteRenderer.cpp:46	D3DFMT_INDEX32	D3DFMT_INDEX16
SpriteRenderer.cpp:111-113	sizeof(uint32_t)	sizeof(uint16_t)
BatchBuilder.h:42	const uint32_t* GetIndices()	const uint16_t* GetIndices()
BatchBuilder.h:54	uint32_t m_indexPool[...]	uint16_t m_indexPool[...]
BatchBuilder.cpp:99	uint32_t* idx	uint16_t* idx
Корневая причина бага: буфер индексов создавался с D3DFMT_INDEX32 (4 байта на индекс), но размер выделялся из расчёта sizeof(WORD) = 2 байта на индекс — ровно вдвое меньше необходимого. memcpy на 32-битных данных писал за границы буфера, коррумпируя последующие аллокации в куче. Ground-ландшафт (первые батчи) читал мусорные индексы, схлопывая полигоны в начало координат

+++ Settlers2/Graphics/BatchBuilder.cpp
@@ -96,7 +96,7 @@ void BatchBuilder::BuildBatches(const RenderCommand* commands, uint32_t commandC
         dst[3].color = cmd.color;
         dst[3].padding[0] = 0; dst[3].padding[1] = 0;
 
-        uint32_t* idx = m_indexPool + m_indexWritePos;
+        uint16_t* idx = m_indexPool + m_indexWritePos;
         idx[0] = baseIdx + 0;
         idx[1] = baseIdx + 1;
         idx[2] = baseIdx + 2;
